Allow user to chose numeral-commands (#2)
The following changes should enable users to choose the numeral and Numeral functions. It took so long because I didn't know how to change the option `use-numerals` which loads 'fmtcount' without breaking something. Luckily, `print-numerals` helps me to remove that problem. Now `use-numerals` doesn't need to be set if numerals want to be used, but the user has to define the numeral-function by him/herself.