Improve error handling for units (and unknown gender)
In the old code, if a not-defined-unit was given it would create an error message, but would still try to execute the next pieces (which rely on a valid unit). This got mostly fixed.

(Note: I am using `\use_none:nnn...` mostly because I am too lazy to indent the code.)