Add commands to set/add/clear unit options
I alway found the key-value approach clunky and a bit annoying. Having commands makes it more manageable; also because you have a better overview.

Furthermore I am adding some basic tests if the units and options exist at all. Bit primitive the code, but I don't expect this to be used by a lot, so otimisation is not really that important here.

Actually quite happy with the result. Note: I don't deprecate the key-value options. Not quite sure what to do with them.