Finish implementation of `master_changer`
The idea behind the `master_changer_prop` is:
Don't evaluate all possible combinations between connected unit during the definition, but only when needed. The evaluation (if needed) is then stored in the global `master_changer_prop` list; the next time the finished value is taken from this list.

Futhermore, the `master_changer_prop` list is cleared everytime a new unit-key is defined or added as in theory the value of a unit could have been changed.

As such, not every unit has a `l_@@_cukeys_ #1 _prop` list, but only the `base unit` (the one used in the first argument of `\cudefin(single)keys`). New units are added to tje `base unit`list (after some conversion).

The main advantage of this method is that the conversion values are only calculated when needed. This should help making the implementation of #14 easier. Furhtermore, I think there isn't really any loss of speed anyway or at least no significant.