Experimental Change: Unit testing
I will ponder about this a bit more. The old system checks if the unit is in a seuence, the new one checks if a variable exists.

As Unit Definitions are global this shouldn't make much difference, but I will check that later.
I could start working with strings, but this would be another seperate version.