New (experimental) function: declarecookingderivative
Currently *not* usable with single-keys.

Example:

    \declarecookingderivative{m,h}{#1:#2}{(#1)/(#2)}{\sfrac{#1}{#2}}

Has a maximum up to 8 units; altough it will take a while. Tried to reduce the bottlencks within my reach as much as possible, but definition of quite a lot of keys can be expensive. That is also the reason why sometimes I wonder if I should try to get rid of keys for this system altogether, but not now.